requirements.txt
torch
transformers
flask
opencv-python
pillow
